rename connection add input name to key to avoid duplicate keys
